+++ sdk/subscription/arm-subscriptions/samples/v7-beta/javascript/subscriptionCancelSample.js
@@ -1,0 +1,24 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { SubscriptionClient } = require("@azure/arm-subscriptions");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to the operation to cancel a subscription
+ *
+ * @summary the operation to cancel a subscription
+ * x-ms-original-file: 2025-11-01-preview/cancelSubscription.json
+ */
+async function cancelSubscription() {
+  const credential = new DefaultAzureCredential();
+  const client = new SubscriptionClient(credential);
+  const result = await client.subscription.cancel("83aa47df-e3e9-49ff-877b-94304bf3d3ad");
+  console.log(result);
+}
+
+async function main() {
+  await cancelSubscription();
+}
+
+main().catch(console.error);